refactor: 기존 코드 제거
