--- src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -1,16 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderLineItem;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface OrderLineItemDao {
-    OrderLineItem save(OrderLineItem entity);
-
-    Optional<OrderLineItem> findById(Long id);
-
-    List<OrderLineItem> findAll();
-
-    List<OrderLineItem> findAllByOrderId(Long orderId);
-}
