--- src/main/java/kitchenpos/dao/OrderTableDao.java
@@ -1,18 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderTable;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface OrderTableDao {
-    OrderTable save(OrderTable entity);
-
-    Optional<OrderTable> findById(Long id);
-
-    List<OrderTable> findAll();
-
-    List<OrderTable> findAllByIdIn(List<Long> ids);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-}
