--- src/main/java/kitchenpos/domain/MenuGroup.java
@@ -1,22 +1,0 @@
-package kitchenpos.domain;
-
-public class MenuGroup {
-    private Long id;
-    private String name;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(final Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-}
