--- src/main/java/kitchenpos/domain/Product.java
@@ -1,33 +1,0 @@
-package kitchenpos.domain;
-
-import java.math.BigDecimal;
-
-public class Product {
-    private Long id;
-    private String name;
-    private BigDecimal price;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(final Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
-    }
-}
