+++ src/main/java/kitchenpos/ordertablegroup/dto/OrderTableGroupCreateRequest.java
@@ -3,8 +3,6 @@ package kitchenpos.ordertablegroup.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
 import kitchenpos.ordertable.dto.OrderTableDto;
 
 public class OrderTableGroupCreateRequest {
@@ -31,18 +29,5 @@ public class OrderTableGroupCreateRequest {
 
 	public void setOrderTables(List<OrderTableDto> orderTables) {
 		this.orderTables = orderTables;
-	}
-
-	public TableGroup toOrderTableGroup() {
-		TableGroup tableGroup = new TableGroup();
-		tableGroup.setOrderTables(
-			orderTables.stream()
-				.map(ot -> {
-					OrderTable orderTable = new OrderTable();
-					orderTable.setId(ot.getId());
-					return orderTable;
-				})
-				.collect(Collectors.toList()));
-		return tableGroup;
 	}
 }
